\name{org.Spombe.egORGANISM}
\alias{org.Spombe.egORGANISM}
\title{The Organism for org.Spombe.eg}
\description{
  org.Spombe.egORGANISM is an R object that contains a single item: a
  character string that names the organism for which org.Spombe.eg was built.
}
\details{
  Although the package name is suggestive of the organism for which it
  was built, org.Spombe.egORGANISM provides a simple way to programmatically
  extract the organism name.
}
\examples{
  org.Spombe.egORGANISM
}
\keyword{datasets}

